Organize data/pybind files
- Reduced nesting of pybind_manager.h
- Renamed runtime bindings to appropriate name
