+++ include/data/pybind/startup_bindings.h
@@ -1,7 +1,7 @@
 // This file is subject to the terms and conditions defined in 'LICENSE' in the source code package
 
-#ifndef JACTORIO_INCLUDE_DATA_PYBIND_PYBIND_BINDINGS_H
-#define JACTORIO_INCLUDE_DATA_PYBIND_PYBIND_BINDINGS_H
+#ifndef JACTORIO_INCLUDE_DATA_PYBIND_STARTUP_BINDINGS_H
+#define JACTORIO_INCLUDE_DATA_PYBIND_STARTUP_BINDINGS_H
 #pragma once
 
 #include <algorithm>
@@ -295,4 +295,4 @@ PYBIND11_EMBEDDED_MODULE(jactorioData, m) {
 #undef PYBIND_PROP_GET_SET
 
 
-#endif // JACTORIO_INCLUDE_DATA_PYBIND_PYBIND_BINDINGS_H
+#endif // JACTORIO_INCLUDE_DATA_PYBIND_STARTUP_BINDINGS_H
