+++ src/CMakeLists.txt
@@ -31,10 +31,9 @@ set(JACTORIO_SRC_FILES
 
         ${JACTORIO_DIR}/data/local_parser.cpp
         ${JACTORIO_DIR}/data/prototype_manager.cpp
+        ${JACTORIO_DIR}/data/pybind_manager.cpp
         ${JACTORIO_DIR}/data/save_game_manager.cpp
         ${JACTORIO_DIR}/data/unique_data_manager.cpp
-
-        ${JACTORIO_DIR}/data/pybind/pybind_manager.cpp
 
 
         ${JACTORIO_DIR}/game/input/input_manager.cpp
